fix(objects,server): complete array classifier + test hardening (PR #272 review)
Review follow-ups on the array-index gate:

- Complete the classifier for the remaining modeled arrays: Command ACTION
  (Table 12-12) and Staging STAGES/STAGE_NAMES/TARGET_REFERENCES
  (Table 12-80) admit an index; ACTION_TEXT stays rejected (array in the
  standard but unmodeled in-tree), as do Channel's EXECUTION_DELAY /
  CONTROL_GROUPS (unmodeled). Correct the citation: the other BACnetLIST
  carrier of LIST_OF_OBJECT_PROPERTY_REFERENCES is Timer (Table 12-75), not
  Lighting Output. Unmodeled array identifiers are documented as rejected
  until their object-side modeling lands.
- Split analog/tests/input_output.rs (707 LOC) into input.rs + output.rs
  under the 700-LOC CI cap.
- Add WPM atomicity proof: a mixed request (valid DESCRIPTION write +
  gated indexed RECIPIENT_LIST write) fails as a whole and applies nothing.
- Pin exact class/code in the Recipient_List legacy-flat arm; add the RP
  wire-path acceptance for Notification Class PRIORITY (BACnetARRAY[3],
  Table 12-24) indexes 0 and 1..=3; rename the Schedule test to the RP+WP
  surface it actually covers.
