+++ crates/bacnet-objects/src/analog/tests/mod.rs
@@ -1,3 +1,4 @@
 mod event_history_read;
-mod input_output;
+mod input;
+mod output;
 mod value;
