+++ crates/bacnet-objects/src/notification_class/tests/recipient_list.rs
@@ -128,7 +128,7 @@ fn recipient_list_indexed_write_rejected_list_unchanged() {
     }
 
     // Legacy flat single-entry shape (what an indexed network write decodes
-    // to) — rejected too.
+    // to) — rejected with the same classification.
     let flat_single = PropertyValue::List(vec![
         PropertyValue::BitString {
             unused_bits: 1,
@@ -144,14 +144,27 @@ fn recipient_list_indexed_write_rejected_list_unchanged() {
             data: vec![0xE0],
         },
     ]);
-    assert!(nc
+    match nc
         .write_property(
             PropertyIdentifier::RECIPIENT_LIST,
             Some(1),
             flat_single,
-            None
+            None,
         )
-        .is_err());
+        .unwrap_err()
+    {
+        Error::Protocol { class, code } => {
+            assert_eq!(
+                class,
+                bacnet_types::enums::ErrorClass::PROPERTY.to_raw() as u32
+            );
+            assert_eq!(
+                code,
+                bacnet_types::enums::ErrorCode::PROPERTY_IS_NOT_AN_ARRAY.to_raw() as u32
+            );
+        }
+        other => panic!("expected PROPERTY/PROPERTY_IS_NOT_AN_ARRAY, got {other:?}"),
+    }
 
     // The whole list is untouched.
     assert_eq!(nc.recipient_list.len(), 3);
